requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
pydantic
pydantic-settings
